Refactor server into modular architecture and fix missing epg_mappings table
