+++ src/middleware/auth.js
@@ -1,0 +1,58 @@
+import jwt from 'jsonwebtoken';
+import rateLimit from 'express-rate-limit';
+import dotenv from 'dotenv';
+
+dotenv.config();
+
+export const JWT_SECRET = process.env.JWT_SECRET || 'change-this-secret-key-in-production';
+export const JWT_EXPIRES_IN = process.env.JWT_EXPIRES_IN || '24h';
+
+// Rate limiting for authentication endpoints
+export const authLimiter = rateLimit({
+  windowMs: 15 * 60 * 1000, // 15 minutes
+  max: 5, // 5 attempts
+  message: { error: 'Too many authentication attempts, please try again later' },
+  standardHeaders: true,
+  legacyHeaders: false,
+});
+
+// General API rate limiting
+export const apiLimiter = rateLimit({
+  windowMs: 1 * 60 * 1000, // 1 minute
+  max: 100, // 100 requests per minute
+  message: { error: 'Too many requests, please try again later' },
+  standardHeaders: true,
+  legacyHeaders: false,
+});
+
+// JWT Authentication Middleware
+export function authenticateToken(req, res, next) {
+  const authHeader = req.headers['authorization'];
+  const token = authHeader && authHeader.split(' ')[1];
+
+  if (!token) {
+    return res.status(401).json({ error: 'No token provided' });
+  }
+
+  jwt.verify(token, JWT_SECRET, (err, user) => {
+    if (err) {
+      return res.status(403).json({ error: 'Invalid or expired token' });
+    }
+    req.user = user;
+    next();
+  });
+}
+
+// Generate JWT token
+export function generateToken(user, isAdmin = false) {
+  return jwt.sign(
+    {
+      userId: user.id,
+      username: user.username,
+      is_active: user.is_active,
+      isAdmin: isAdmin
+    },
+    JWT_SECRET,
+    { expiresIn: JWT_EXPIRES_IN }
+  );
+}
