+++ src/utils/helpers.js
@@ -1,0 +1,35 @@
+import { Xtream } from '@iptv/xtream-api';
+
+export function isAdultCategory(name) {
+  const adultKeywords = [
+    '18+', 'adult', 'xxx', 'porn', 'erotic', 'sex', 'nsfw',
+    'for adults', 'erwachsene', '+18', '18 plus', 'mature',
+    'xxx', 'sexy', 'hot'
+  ];
+  const nameLower = name.toLowerCase();
+  return adultKeywords.some(kw => nameLower.includes(kw));
+}
+
+export function calculateNextSync(interval) {
+  const now = Math.floor(Date.now() / 1000);
+  switch (interval) {
+    case 'hourly': return now + 3600;
+    case 'every_6_hours': return now + 21600;
+    case 'every_12_hours': return now + 43200;
+    case 'daily': return now + 86400;
+    case 'weekly': return now + 604800;
+    default: return now + 86400;
+  }
+}
+
+export async function cachePicon(originalUrl, channelName) {
+  // Simply return the original URL - no caching needed
+  return originalUrl || null;
+}
+
+export function createXtreamClient(provider) {
+  let baseUrl = (provider.url || '').trim();
+  if (!/^https?:\/\//i.test(baseUrl)) baseUrl = 'http://' + baseUrl;
+  baseUrl = baseUrl.replace(/\/+$/, '');
+  return new Xtream({ url: baseUrl, username: provider.username, password: provider.password });
+}
